requirements.txt
transformers==4.31.0
speechtokenizer==0.1.2
tokenizers==0.13.3
datasets==2.14.6
nlp2==1.8.52
soundfile==0.12.1
librosa==0.10.1
pesq==0.0.4
descript-audio-codec
descript-audiotools